feat: add 'endGame' function to display end game message to the victorious
add endgame html template message displaying the correct emoji and message. Using ternary operator for the logic, this function checks both wizard's and orc's health to display the right message. 'endGame' is called inside the 'attack' function and before is called, it checks with an if statement if 'wizard' or the 'orc' dead boolean is truthy or falsy